refactor: move into own package and rename quiz summary fragment
